[iris] Drop ControllerConfig.finelog; fifth lint-review pass
ControllerConfig.finelog lost its only reader when the relay was deleted; main.py
still plumbed it into a field nothing consumed. The live reader of the cluster
config's finelog block (deriving /system/log-server) is untouched. cluster_id no
longer claims to stamp relayed finelog batches -- the finelog server stamps rows,
from the key that authenticated the push.

- unique_dir moves to a crate-root test_support, so query's tests no longer reach
  sideways into server's fixtures for a temp-dir helper.
- wait_for_cursor and wait_for_requests share one poll_until scaffold.
- resume_after_eviction's doc claimed None means "cursor at or above the oldest
  local row"; the guard is cursor + 1 < min_seq, so a cursor one below min_seq is
  also None -- the next row is still there.
