+++ lib/iris/src/iris/cluster/controller/main.py
@@ -279,7 +279,6 @@ def run_controller_serve(
         autoscaler_evaluation_interval=cluster_config.defaults.autoscaler.evaluation_interval,
         cluster_id=cluster_config.name,
         peers=cluster_config.peers,
-        finelog=cluster_config.finelog,
     )
 
     # Each worker-daemon backend constructs and owns its liveness tracker, sized by
